reorganized the Search component into 2 files
