+++ src/content/utils.ts
@@ -1,4 +1,4 @@
-import { Action, Message, TabData } from "../common/types";
+import { Message, TabData } from "../common/types";
 import browser from "webextension-polyfill";
 
 export function getCurrentTabData() {
@@ -13,24 +13,3 @@ export function isInvalidatedContextError(error: Error) {
 }
 
 
-export const filterByCurrentWindow = (currentTabs: TabData[]) => {
-  return currentTabs.filter((tabData) => tabData.inCurrentWindow);
-};
-
-
-export const tabMatchesValue = (searchValue: string, tabData: TabData) =>
-  tabData.tabTitle.toLowerCase().includes(searchValue.toLowerCase()) ||
-  tabData.tabUrl.toLowerCase().includes(searchValue.toLowerCase());
-
-export const filterTabs = (searchValue: string, currentTabs: TabData[]) => {
-  return currentTabs.filter(
-    (tabData) => tabMatchesValue(searchValue, tabData),
-    // try to filter based on the tab title and the tab url
-  );
-};
-
-export const filterActions = (searchValue: string, actions: Action[]) => {
-  return actions.filter((action) =>
-    action.name.toLowerCase().includes(searchValue.toLowerCase()),
-  );
-};
